Document that `Send + Sync` also effectively applies to `BORROWED`.
This means that `NonSync` will be used more frequently. Considered
shortening it to `N`, but since people are more likely to encounter the
limitation the hard way and then try to scan the docs, `NonSync` makes
it obviously related and they can rename-use the symbol to something
shorter if they like.